Move `CurrencySelect` to an individual tsx file to improve readability
